Display the bill in terminal and separate input taking part from the cart class
